Make skirt/brim line width and flow visible for draft/ooze shields
The draft and ooze shield use the skirt/brim settings to print. So these settings should be adjustable if the draft or ooze shield is enabled. It's a bit weird UX-wise, but it's correct. The UX problem will be tackled later.
Apparently this was already done for the speed/acceleration/jerk settings so no change necessary there.

Contributes to issue #8808.