Fix support density when using tree support on engineering materials
These formulas were using a setting that didn't exist any more.
Not sure how this got through QA. Oh well, this is what they are intending.
This formula is a little bit weird in how it works with support meshes if support is disabled, but it's more or less as good as it gets. The formula is mirrored from how other Ultimaker printers write it down, but slightly simplified with the same outcome.