Connect the visibility of the components through their properties
As Ghostkeeper suspected correctly in the review comment https://github.com/Ultimaker/Cura/pull/9012#discussion_r549707433
the binding wasn't working because the model was being retrieved using a function
(CuraApplication.getFileProviderModel()).

Separating this model into a variable allows us to properly bind the "visible" properties of the
menu items with the count property of the model without a problem.

CURA-7868