Move open dialogue to separate file provider plug-in
We can now define plug-ins that specify where to open files from. This is one of the places where you can open files.
This breaks the main button to open files in the interface. It needs to be redirected to trigger the plug-in to show the open file dialogue.

Contributest o issue CURA-7868.