Display the correct message when opening deleted files from Recent
If a file that existed in the Recent Files list was deleted, Cura was still trying to open the file
on user's request without properly displaying an error if the file was not found.

This is now fixed by popping up a message to indicate that the file doesn't exist.

CURA-7996