Add 'Outer Wall Lock Factor' setting.
Arachne produced artifacts in straight outer walls if the inner walls had curves in them (or appeared and dissapeared behind them). 'Hard' fixing that would reduce the flexibility of Arachne, so a setting is introduced how much the outer walls should be locked into place (as much as possible at 100%, not at all at 0%).

CURA-7979