fix: Removes custom setting in machine def file
Removes custom settings in the machine definition files. These were used to set custom values for the maximum print temperature, maximum retraction speed and distance based on the tool which is used (PTFE or full metal hot end). Thus all variants are also updated. This is done based on feedback of a pull request.