Handle deleting profiles better
When deleting profiles, the current item is set to null.
The Activate button was checking if the current profile was already active, and disabling the button if it was. But the button was still enabled if the current item is null because isCurrentItemActivated is only true if there is a current item. So to properly disable the button we need to check also if there is a current item.
The onItemsChanged signal can also trigger if the item that matches the toSelectItemName has no quality_changes_group, i.e. a built-in profile.

Probably fixes Sentry issue CURA-43.