Add a separate stage for drawing the alpha component for transparent parts of the render
Transparent components need to be drawn on top of solid components.
We use the geometry shader in order to separate out the transparent parts from the solid parts, because they are intermixed within one vertex buffer. In effect this doesn't significantly increase the render time because all vertices and fragments are still drawn only once. But there is quite a bit more overhead now in that it still does check twice as much which parts need to be rendered in this pass (execute the geometry shader twice). And then it needs to blend two extra render batches into the final composite. The two extra render batches are not really going to be a problem (there's already a dozen in the scene) but the geometry shader is executed twice, really, due to GPUs not being able to properly branch the execution.