Create MultilayerBrim.py
This plugin is meant to encounter warping, by increasing the adhesion of the brim to the heated bed. It does so by dividing the single brim layer into multilayers of smaller height. It respects the original height of the brim and filament extruded, and divides it over the configurable number of layers.