Hard quit. (Instead of closing gently.)
As the comments for this method already state, anything of value should already have been saved when calling this. It may not be the prettiest workaround, but why bother cleaning things up before they go in the fires of /dev/nul anyway? Now the question is ... will it work in the build?

CURA-7813