Slightly lower build plate mesh
The top of the build plate mesh seems to be 1mm into the build volume in its actual coordinates. The mesh was lowered by 0.8mm. Probably they tried to lower it by 1mm first and then found that it was z-fighting. I'm resolving the z-fight the other way around so that the mesh actually appears completely below the build volume. This makes the build plate grid more clear. No more grey film on top of it.