Hide initial layer speed settings if Number of Slower Layers is 0
These initial layer speeds were only applied to the model (not the raft).
Number of Slower Layers indicates how many layers the initial layer speed should apply to. So if that is 0, the initial layer speed doesn't apply to anything any more, and the setting should be hidden because it has no effect on the print.

Found by Markouustalu in #4439.