Catch environment errors when restoring back-ups
There could be an environment error when saving that file because of access rights or insufficient disk space. Catch that error and display an error message to the user. The log then contains more information on exactly why it failed, but the user just knows it fails.

Fixes Sentry issue CURA-21W.