Reduce gantry height of UMS5 and UMS3
One of our engineers found experimentally that his 60mm model was crashing into the gantry. And indeed it is confirmed with the designs of the UM3/S5 print head that the space between the nozzle tip and bottom of the gantry is 57mm. Allowing for some play in the dimensions of the print head and the positioning of the nozzle in the print head, 55mm is agreed as a safe distance.

Fixes issue CURA-8182.