Fix getting PyCapsule error on import
Previously, Arcus, Savitar, and pynest2d were being build with a missing sip flag. This led to Cura
getting `ValueError: PyCapsule_GetPointer called with incorrect name`, if those libraries were not
imported (even if they were unused).

This flag was added in the following commits:

* https://github.com/Ultimaker/libSavitar/commit/83393f989c0a8eb0b4ddc00560f028bf4eb4ac6b
* https://github.com/Ultimaker/libArcus/commit/e9e530d96383c18a739e7ad181907c33b30157ec
* https://github.com/Ultimaker/pynest2d/commit/bd440b50ad6fe969063ecafe8ad99a7b7a25ccc2

Since the issue seems to be gone now, the unused imports should be removed from Cura.

All credits go to Rex Dieter for figuring that this missing flag was the issue.
https://src.fedoraproject.org/rpms/libarcus/pull-request/1#request_diff

CURA-8146