Keep Z Hop Speed visible
It's also used as the speed at which layer changes are made, so technically it should be kept visible since the setting does indeed affect the g-code.
In practice, it won't affect the print though. Using a reasonable Z hop speed of 10mm/s and the default acceleration of 100mm/s/s, it would take 0.1s to reach the Z hop speed. With the formula for acceleration, 1/2att+vt+c, you can see that it needs 0.5mm to get up to speed, and another 0.5mm to get back to 0. So this speed is only attained with layer heights of more than 1mm which is very unrealistic. For this reason we won't document this in the setting description or its name. The Z hop speed is only used for layer changes because it needs a sensible default for extreme cases, not because it needs to be configurable as such.

But we'll keep it visible to give a less-than-expert user a feeling of control, a placebo effect that they are actually making print quality improvements even if their firmware forbids it from making an actual difference.

Fixes Ultimaker/CuraEngine#1446.