Check if there is a global stack when discard/keep dialog closes
When we close Cura, we now remove the global stack first. If the discard/keep dialog closes as a result of closing Cura, this would crash since there is no global stack any more to discard changes from or keep changes in, and we can't access the extruder stacks either.
So if there's no global stack, just skip this. It'll be as if you selected 'keep'.

Fixes Sentry issue CURA-2ET.