Show warning when restoring backup failed
We now allow the backup to fail gracefully when partial files
fail to be restored. But the user is not actively informed by
these failures. Leaving him/her unaware of the state of his
configuration folder.

This commit show's a message windows, listing the failed files.
Due to a string freeze at the moment we reuse the following i18n
message:

> The following error occurred while trying to restore a Cura backup:
followed by a list of files that fails to be restored.

Which is not ideal, but a sufficient warning in my opinion. At least
better then being completely uninformed.

Contributes to CURA-8313