Change the color of a circle around the toolbar button
Instead of changing the color of the entire tool when it is hovered or selected, now a background circle will be drawn and its color will change according to whether the button is hovered/selected or not.

In addition, the button icons will now be half the size of the button itself.

CURA-8013