Replace the shadow of the PrepareMenu with a border
In order for that to be possible, the background border of the ExpandablePopup and the ExpandableComponent had to be exposed.

In addition, since in the `RowLayout` the `ConfigurationMenu` and the `MachineSelector` both draw a border around them, it means that there will be a double "divider" (one border line from each button). To overcome that, we set the spacing of the `RowLayout` as a `- border_width`, which ensures then that there will be only a single line of border between the two adjacent buttons.

CURA-8013