Rework parts of camera button
The new button should be consistent with the looks of the tool buttons. I've given it the same size as the inner part of the tool buttons, reworked the background so that it matches. I'm also giving it the 'secondary button' style which is also used for one other button. But that other detail button should get the same style anyway so that's fine too.

I've also turned this icon into an actual button. That way it can get focus when using keyboard navigation and works better with tools for blind and such.

Contributes to issue CURA-8202.