Fix margins
The margin around the icon didn't work if the icon source was none, because it was provided a QSize rather than a real.
The margin around the background didn't work at all because the background is automatically adjusted to fill the parent element, so it's best left out.

Contributes to issue CURA-8009.