Stop slicing if an error occurred
If the user has auto-slicing enabled, this causes the indeterminate progress bar to appear.
If the user doesn't have auto-slicing enabled, this causes the slice button to reappear.

Both of these indicate that something has caused slicing to be interrupted. The message should make clear why.

Contributes to issue CURA-6568.