Give secondary buttons a background colour again
This background colour is the same as the main background, on which these buttons almost always are situated. However in the cases where they are not, such as the post-processing scripts button, this makes the text easier to read than a completely transparent background colour.

Done as a 5 minute fix.