Fix long material names overlapping with extruder icons
Both the `preferredWidth` AND the `width` need to be set for the labels to be properly elided if there is no more room  in the ColumnLayout. In addition, the ColumnLayout that contains the variants and material names needs to be visible only if width>0, otherwise when the width is negative, for some reason the `materialTypeLabel` appears again.

CURA-8496