Slightly shorten AnyCubic i3 Mega S/Pro name
This was causing some files to be just over the limit of file length when given the default printer name on some operating systems. Some file systems, especially encrypted ones, have strict limitations for the maximum length of a file name. With percent-encoding and the extension and postfix we add to these file names, this could go over the limit in file name length for some operating systems.
Easy fix is to slightly shorten the name, at least to remove those brackets.

Done as a 5 minute fix.