Change the ColumnLayout to Column in the ConfigurationMenu
When the labels were getting truncated and invisible after reducing the width of Cura's window, their visibility wasn't being restored back and, as a result, the `materialTypeLabel` was the only one that was remaining visible, even if there was enough space for the full `materialBrandColorTypeLabel`.

Changing the ColumnLayout to a Column, where the width is inherited from the parent, fixes that issue.

CURA-8496