Use custom style for the links in the what's new textArea
A style had to be prepended in the text to make sure that it adheres to the style we want instead of using the OS-based link style.

Note: For some reason QML ignores all other link-style tags (such as `a:hover`, `a:focus` etc) apart from `a:link`, so we can only change the style for the normal link and NOT for when it is clicked or hovered.

CURA-8380